fix: refactor app context and app file
