+++ src/context/authAndPortalDataProvider/index.tsx
@@ -1,134 +1,21 @@
-import { createContext, useEffect, useMemo, useState } from "react";
-import { useAuth0 } from "@auth0/auth0-react";
-
-import { IBusinessUnitsPortalStaff } from "@ptypes/staffPortal/IBusinessUnitsPortalStaff";
-import { decrypt } from "@utils/crypto/decrypt";
-
-import { usePortalData } from "@hooks/staffPortal/usePortalData";
-import { useBusinessManagers } from "@hooks/staffPortal/useBusinessManagers";
+import { createContext } from "react";
 import { IAuthAndPortalData } from "@ptypes/context/authAndPortalDataProvider/IAuthAndPortalData";
-import { IAppData } from "@ptypes/context/authAndPortalDataProvider/IAppData";
+import { useValidatingLoginInformation } from "@hooks/useValidatingLoginInformation";
 
-
-const AuthAndPortalData = createContext<IAuthAndPortalData>({} as IAuthAndPortalData);
-
+const AuthAndPortalData = createContext<IAuthAndPortalData>(
+  {} as IAuthAndPortalData,
+);
 interface IAuthAndPortalDataProvider {
   children: React.ReactNode;
 }
 
 function AuthAndPortalDataProvider(props: IAuthAndPortalDataProvider) {
   const { children } = props;
-  const { user } = useAuth0();
-
-  const portalCode = decrypt(localStorage.getItem("portalCode") ?? "");
-  const { portalData } = usePortalData(portalCode);
-  const { businessManagersData } = useBusinessManagers(portalData);
-  const [businessUnitSigla, setBusinessUnitSigla] = useState(
-    localStorage.getItem("businessUnitSigla") ?? "",
-  );
-  const [businessUnitsToTheStaff, setBusinessUnitsToTheStaff] = useState<
-    IBusinessUnitsPortalStaff[]
-  >(() => {
-    const savedBusinessUnits = localStorage.getItem("businessUnitsToTheStaff");
-    return savedBusinessUnits ? JSON.parse(savedBusinessUnits) : [];
-  });
-
-  let businessUnitData: IBusinessUnitsPortalStaff =
-    {} as IBusinessUnitsPortalStaff;
-  try {
-    businessUnitData = JSON.parse(
-      businessUnitSigla || "{}",
-    ) as IBusinessUnitsPortalStaff;
-  } catch (error) {
-    console.error("Error parsing businessUnitSigla:", error);
-  }
-  const [appData, setAppData] = useState<IAppData>({
-    portal: {
-      abbreviatedName: "",
-      staffPortalCatalogId: "",
-      businessManagerId: "",
-      publicCode: "",
-    },
-    businessManager: {
-      publicCode: "",
-      abbreviatedName: "",
-      urlBrand: "",
-      urlLogo: "",
-    },
-    businessUnit: {
-      publicCode: businessUnitData?.publicCode || "",
-      abbreviatedName: businessUnitData?.abbreviatedName || "",
-      languageId: businessUnitData?.languageId || "",
-      urlLogo: businessUnitData?.urlLogo || "",
-    },
-    user: {
-      userAccount: user?.email ?? "",
-      userName: user?.name ?? "",
-    },
-  });
-
-  useEffect(() => {
-    if (!businessManagersData) return;
-
-    setAppData((prev) => ({
-      ...prev,
-      portal: {
-        ...prev.portal,
-        abbreviatedName: portalData?.abbreviatedName || "",
-        staffPortalCatalogId: portalData?.staffPortalId || "",
-        businessManagerId: portalData?.businessManagerId || "",
-        publicCode: portalData?.publicCode || "",
-      },
-      businessManager: {
-        ...prev.businessManager,
-        publicCode: businessManagersData.publicCode || "",
-        abbreviatedName: businessManagersData.abbreviatedName || "",
-        urlBrand: businessManagersData.urlBrand || "",
-        urlLogo: businessManagersData.urlLogo || "",
-      },
-    }));
-  }, [businessManagersData, portalData, portalCode]);
-
-  useEffect(() => {
-    localStorage.setItem("businessUnitSigla", businessUnitSigla);
-
-    if (businessUnitsToTheStaff && businessUnitSigla) {
-      const businessUnit = JSON.parse(businessUnitSigla);
-
-      setAppData((prev) => ({
-        ...prev,
-        businessUnit: {
-          ...prev.businessUnit,
-          abbreviatedName: businessUnit?.abbreviatedName,
-          publicCode: businessUnit?.publicCode,
-          languageId: businessUnit?.languageId,
-          urlLogo: businessUnit?.urlLogo,
-        },
-      }));
-    }
-  }, [businessUnitSigla, businessUnitsToTheStaff]);
-
-  useEffect(() => {
-    localStorage.setItem(
-      "businessUnitsToTheStaff",
-      JSON.stringify(businessUnitsToTheStaff),
-    );
-  }, [businessUnitsToTheStaff]);
-
-  const authAndPortalDataContainer = useMemo(
-    () => ({
-      appData,
-      businessUnitSigla,
-      businessUnitsToTheStaff,
-      setAppData,
-      setBusinessUnitSigla,
-      setBusinessUnitsToTheStaff,
-    }),
-    [appData, businessUnitSigla, businessUnitsToTheStaff],
-  );
-
+  const { authAndPortalDataContainer } = useValidatingLoginInformation();
   return (
-    <AuthAndPortalData.Provider value={authAndPortalDataContainer}>{children}</AuthAndPortalData.Provider>
+    <AuthAndPortalData.Provider value={authAndPortalDataContainer}>
+      {children}
+    </AuthAndPortalData.Provider>
   );
 }
